Move autocomplete tests to own file
